cloud/backend/vehicle_gateway/pyproject.toml
[tool.poetry]
name = "vehicle-gateway"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
eclipse-zenoh = "1.0.0"
protobuf = "==3.20.*"
influxdb-client = "1.45.0"
fastapi = { extras = ["standard"], version = "^0.115.0" }
json5 = "0.9.25"
pymongo = "<4.10"
motor = "^3.6.0"
httpx = "^0.27.2"
google-api-python-client = "^2.161.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"

[tool.poetry.scripts]
format = "scripts:format_code"
lint = "scripts:lint_code"
proto = "scripts:generate_proto"
dev = "scripts:run_dev"
prod = "scripts:run_prod"

[tool.black]
line-length = 88
target-version = ['py310']
exclude = '''
/(
    venv
    | __pycache__
    | vehicle_gateway/generated
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
